Only try and delete files that exist
I did use AI to figure out how to delete files with a glob that may or
may not exist. It's not Erlang, so I'm fine with it. I did figure out my
own way with a for loop and a file check, but I don't think I want to be
a bash expert anymore.

+++ t
@@ -2,7 +2,14 @@
 
 rm -rf logs/
 
-rm test/level_*.beam
+# create array, check if 1st element is a file
+# if test/level_*.beam doesn't match anything the first element
+# of the array will be the string "test/level_*.beam"; i.e. the shell
+# won't expand the value if no files match
+files=(test/level_*.beam)
+if [[ -e ${files[0]} ]]; then
+  rm test/level_*.beam
+fi
 
 touch src/*
 touch test/*
